Move MultipleFileField to custom_forms.py
