+++ omeroweb/webclient/custom_forms.py
@@ -379,3 +379,23 @@ class ObjectModelMultipleChoiceField(ObjectModelChoiceField):
                 else:
                     final_values.append(val)
         return final_values
+
+
+# Custom widget and validation for multiple file uploads
+# See https://docs.djangoproject.com/en/3.2/topics/http/file-uploads/#uploading-multiple-files
+class MultipleFileInput(forms.ClearableFileInput):
+    allow_multiple_selected = True
+
+
+class MultipleFileField(forms.FileField):
+    def __init__(self, *args, **kwargs):
+        kwargs.setdefault("widget", MultipleFileInput())
+        super().__init__(*args, **kwargs)
+
+    def clean(self, data, initial=None):
+        single_file_clean = super().clean
+        if isinstance(data, (list, tuple)):
+            result = [single_file_clean(d, initial) for d in data]
+        else:
+            result = single_file_clean(data, initial)
+        return result
